flask-app/requirements.txt
Flask==3.1.0
flask-cors==5.0.1
matplotlib==3.10.1
numpy==2.2.3
pynmea2==1.19.0
pyserial==3.5
requests==2.32.3
smbus==1.1.post2
